pyproject.toml
[project]
name = "cinema"
version = "0.1.0"
description = ""
authors = [
    { name = "Mohammadreza H.Rezaeian", email = "mdreza.rzn@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic[email] (<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "databases[sqlite] (>=0.9.0,<0.10.0)",
]

[tool.poetry.scripts]
app = "app.cinema:create_run_app"
import_mock_data = "mock.import_all:main"

[tool.poetry]
packages = [{ include = "app", from = "" }]



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
